Because the etherscan feature will be removed from the plugin at some point, dev for this (issue #27) and the etherscan related feature will be paused.
The plugin takes etherscan's api key and generates a button and url for a json response to etherscan's ethprice action.

I can not test etherscan&orders with the current version of the plugin because the version's gateway doesn't work.

The plugin should also use eitherscan for website price and conversion values but this is untested.